[aptos-cli] Cleanup utils & types
